refactor 1/x
